refactor: move modal managers to /ui/managers
