+++ src/main.ts
@@ -1,5 +1,5 @@
 import { Plugin } from "obsidian";
-import { XoppSettingsTab } from "src/XoppSettingsTab";
+import { XoppSettingsTab } from "src/ui/modals/xopp-settings-tab";
 import { checkXoppSetup } from "src/core/environment-checks";
 import { createCommands } from "src/core/commands";
 import { setupListeners } from "src/core/listeners";
