+++ src/ribbonIcons.ts
@@ -1,5 +1,5 @@
 import XoppPlugin from "src/main";
-import CreateXoppModalManager from "./CreateXoppModalManager";
+import CreateXoppModalManager from "src/ui/managers/create-xopp-modal-manager";
 
 export function createRibbonIcons(plugin: XoppPlugin) {
     plugin.addRibbonIcon("pen-tool", "Create new Xournal++ note", () => {
