+++ tsconfig.json
@@ -11,6 +11,8 @@
         "importHelpers": true,
         "isolatedModules": true,
         "strictNullChecks": true,
+        "esModuleInterop": true,
+        "skipLibCheck": true,
         "lib": ["DOM", "ES5", "ES6", "ES7"]
     },
     "include": ["**/*.ts"]
